+++ tests/sdk/get_test.go
@@ -2,6 +2,9 @@ package sdk
 
 import (
 	"fmt"
+	"io/ioutil"
+	"math/rand"
+	"testing"
 	"github.com/starkinfra/core-go/starkcore/utils/api"
 	Boleto "github.com/starkinfra/core-go/tests/utils/boleto"
 	Log "github.com/starkinfra/core-go/tests/utils/boleto/log"
@@ -10,9 +13,6 @@ import (
 	"github.com/starkinfra/core-go/tests/utils/issuing/product"
 	"github.com/starkinfra/core-go/tests/utils/sign"
 	User "github.com/starkinfra/core-go/tests/utils/user"
-	"io/ioutil"
-	"math/rand"
-	"testing"
 )
 
 func TestBoletoGet(t *testing.T) {
@@ -32,9 +32,19 @@ func TestBoletoQuery(t *testing.T) {
 	var params = map[string]interface{}{}
 	params["limit"] = 200
 
-	boletos := Boleto.Query(params)
-	for boleto := range boletos {
-		fmt.Println("id:", boleto.Id)
+	boletos, err := Boleto.Query(params)
+	for {
+		select {
+		case errors := <-err:
+			if errors.Errors != nil {
+				for _, e := range errors.Errors {
+					panic(fmt.Sprintf("code: %s, message: %s", e.Code, e.Message))
+				}
+			}
+			return
+		case boleto := <-boletos:
+			fmt.Println(boleto.Id)
+		}
 	}
 }
 
@@ -82,19 +92,33 @@ func TestBoletoLogQuery(t *testing.T) {
 	params["limit"] = 300
 	params["after"] = "2022-11-16"
 
-	boletos := Log.Query(nil)
-
-	for boleto := range boletos {
-		fmt.Println("i, boleto", boleto.Id)
+	boletos, err := Log.Query(params)
+	for {
+		select {
+		case errors := <-err:
+			for _, e := range errors.Errors {
+				panic(fmt.Sprintf("code: %s, message: %s", e.Code, e.Message))
+			}
+			return
+		case boleto := <-boletos:
+			fmt.Println(boleto.Id)
+		}
 	}
 }
 
 func TestProductLogQuery(t *testing.T) {
 
-	products := product.Query(nil)
-
-	for product := range products {
-		fmt.Println("i: product", product.Id)
+	products, err := product.Query(nil)
+	for {
+		select {
+		case errors := <-err:
+			for _, e := range errors.Errors {
+				panic(fmt.Sprintf("code: %s, message: %s", e.Code, e.Message))
+			}
+			return
+		case product := <-products:
+			fmt.Println(product.Id)
+		}
 	}
 }
 
@@ -110,7 +134,7 @@ func TestBoletoLogPage(t *testing.T) {
 		}
 	}
 	for _, boleto := range boletos {
-		fmt.Println(boleto)
+		fmt.Println(boleto.Id)
 	}
 	fmt.Println(cursor)
 }
@@ -169,9 +193,17 @@ func TestWorkspaceReplaceQuery(t *testing.T) {
 	var params = map[string]interface{}{}
 	params["limit"] = 1
 
-	invoices := Invoice.Query(params, User.ExampleOrganization.Replace("4690697751887872"))
-	for invoice := range invoices {
-		fmt.Println("invoice's id: ", invoice.Id)
+	invoices, err := Invoice.Query(params, User.ExampleOrganization.Replace("4690697751887872"))
+	for {
+		select {
+		case errors := <-err:
+			for _, e := range errors.Errors {
+				panic(fmt.Sprintf("code: %s, message: %s", e.Code, e.Message))
+			}
+			return
+		case invoice := <-invoices:
+			fmt.Println(invoice.Id)
+		}
 	}
 }
 
@@ -180,9 +212,16 @@ func TestEventQuery(t *testing.T) {
 	params["isDelivered"] = true
 	params["limit"] = 300
 
-	events := Event.Query(params)
-
-	for event := range events {
-		fmt.Println(event.Id)
+	events, err := Event.Query(params)
+	for {
+		select {
+		case errors := <-err:
+			for _, e := range errors.Errors {
+				panic(fmt.Sprintf("code: %s, message: %s", e.Code, e.Message))
+			}
+			return
+		case result := <-events:
+			fmt.Println(result.Id)
+		}
 	}
 }